requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
plotly>=5.0.0
akshare>=1.9.0
scikit-learn>=1.0.0
talib-binary>=0.4.0
seaborn>=0.11.0
tqdm>=4.62.0
joblib>=1.1.0
